Added classes
Server package

FileReceiverServer
Class that acts as a server which will receive the files (locally).

FileReceiverServerIce
does the same as FileReceiverServerIce class but using p2p + ICE protocol (in process)

Client package

FileSenderClient
class that acts as client that sends files (locally)

FileSenderClientIce
does the same as FileSenderClient but using p2p + ICE method to send files (in process)

Enums package
Protocol
It is an enum containing UDP and TCP protocols.

Records package

Candidate
Data structure representing a connection candidate on the server side FileReceiverServerIce

FileMetadata
Data structure representing the information of a file which is sent before the original file

ServerKeyInfo
Data structure that represents the necessary information from the server Ice(FileReceiverServerIce) for the client(FileSenderClientIce) to make the connection
Allows to export in xml format to send it to the client externally (a secure encryption form will be implemented in the future)

Translated with DeepL.com (free version)